refactor(lib/api): consolidate functions into `db` directory (#227)
* fix(api/routes/people): remove unnecessary API route

* refactor(api/routes): remove plurality from dir names

* refactor(lib/api): move `create/` files to `db/`

* refactor(lib/api): copy `delete/` functions to `db/`

* wip(lib/api): move `get/` functions to `db/`

* wip(lib/api): move `update/` functions to `db/`

* wip(lib/api/db): remove duplicate imports and exports

* wip(lib/api/db): export defined `DB` types
