+++ lib/api/auth.ts
@@ -1,3 +1,46 @@
+import to from 'await-to-js';
+
+import { FirebaseError, auth } from 'lib/api/firebase';
+import { APIError } from 'lib/api/error';
+
+/**
+ * Deletes the Firebase Authentication account for the given user.
+ * @param uid - The user ID whose account we want to delete.
+ * @return Promise that resolves to nothing; throws an `APIError` if we were
+ * unable to delete the Firebase Authentication account.
+ */
+export default async function deleteAuthUser(uid: string): Promise<void> {
+  const [err] = await to<void, FirebaseError>(auth.deleteUser(uid));
+  if (err) {
+    const msg = `${err.name} (${err.code}) deleting auth account (${uid})`;
+    throw new APIError(`${msg}: ${err.message}`, 500);
+  }
+}
+
+import to from 'await-to-js';
+
+import { FirebaseError, auth } from 'lib/api/firebase';
+import { APIError } from 'lib/api/error';
+
+/**
+ * Creates a custom login token that can be used by the client when they signup
+ * with the deprecated org signup page.
+ * @param uid - The user ID of the user to create the login token for.
+ * @return Promise that resolves with the token string; throws an `APIError` if
+ * we were unable to create a custom token.
+ * @todo Remove this once we fix #116 and get rid of that signup page.
+ */
+export default async function createCustomToken(uid: string): Promise<string> {
+  const [err, token] = await to<string, FirebaseError>(
+    auth.createCustomToken(uid)
+  );
+  if (err) {
+    const msg = `${err.name} (${err.code}) creating custom auth token`;
+    throw new APIError(`${msg} for user (${uid}): ${err.message}`, 500);
+  }
+  return token as string;
+}
+
 import phone from 'phone';
 import to from 'await-to-js';
 import { v4 as uuid } from 'uuid';
