+++ lib/api/routes/match/delete.ts
