+++ lib/api/routes/meeting/delete.ts
