+++ lib/api/routes/user/delete.ts
